+++ tests/test_logging.py
@@ -10,13 +10,13 @@ import slash
 from .utils import run_tests_assert_success, run_tests_in_session, TestCase
 
 
-def test_last_session_symlinks(logs_dir):
-    session = run_tests_assert_success(SampleTest)
+def test_last_session_symlinks(logs_dir, session):
 
     files_dir = logs_dir.join("logs", "files")
     links_dir = logs_dir.join("logs", "links")
 
-    test_log_file = files_dir.join(session.id, list(session.results.iter_test_results())[-1].test_metadata.id, "log")
+    test_log_file = files_dir.join(
+        session.id, list(session.results.iter_test_results())[-1].test_metadata.id, "log")
     assert test_log_file.check()
     session_log_file = files_dir.join("session.log")
     assert session_log_file.check()
@@ -25,11 +25,26 @@ def test_last_session_symlinks(logs_dir):
     assert links_dir.join("last-test").readlink() == test_log_file
 
 
+def test_result_log_links(logs_dir, session):
+
+    for result in session.results.iter_test_results():
+        assert result.get_log_path() is not None
+        assert result.get_log_path().startswith(str(logs_dir.join("logs", "files")))
+
+
+@pytest.fixture
+def session():
+    session = run_tests_assert_success(SampleTest)
+    return session
+
+
 @pytest.fixture
 def logs_dir(request, config_override, tmpdir):
     config_override("log.root", str(tmpdir.join("logs", "files")))
-    config_override("log.last_session_symlink", str(tmpdir.join("logs", "links", "last-session")))
-    config_override("log.last_test_symlink", str(tmpdir.join("logs", "links", "last-test")))
+    config_override("log.last_session_symlink",
+                    str(tmpdir.join("logs", "links", "last-session")))
+    config_override("log.last_test_symlink",
+                    str(tmpdir.join("logs", "links", "last-test")))
     return tmpdir
 
 
@@ -38,6 +53,7 @@ _SESSION_START_MARK = "session-start-mark"
 _SESSION_END_MARK = "session-end-mark"
 
 _silenced_logger = logbook.Logger("silenced_logger")
+
 
 class LogFormattingTest(TestCase):
 
@@ -59,7 +75,9 @@ class LogFormattingTest(TestCase):
                 self.assertTrue(line.startswith("-- "))
                 self.assertTrue(line.endswith(" --\n"))
 
+
 class LoggingTest(TestCase):
+
     def test(self):
         self.log_path = self.get_new_path()
         self.override_config(
@@ -68,7 +86,8 @@ class LoggingTest(TestCase):
         )
         self.override_config(
             "log.subpath",
-            os.path.join("{context.session.id}", "{context.test_id}", "debug.log")
+            os.path.join("{context.session.id}",
+                         "{context.test_id}", "debug.log")
         )
         self.override_config(
             "log.session_subpath",
@@ -80,13 +99,16 @@ class LoggingTest(TestCase):
         )
 
         self.addCleanup(gossip.unregister_token, _TOKEN)
-        slash.hooks.session_start.register(functools.partial(_mark, _SESSION_START_MARK), token=_TOKEN)
+        slash.hooks.session_start.register(
+            functools.partial(_mark, _SESSION_START_MARK), token=_TOKEN)
 
-        slash.hooks.session_end.register(functools.partial(_mark, _SESSION_END_MARK), token=_TOKEN)
+        slash.hooks.session_end.register(
+            functools.partial(_mark, _SESSION_END_MARK), token=_TOKEN)
         self.addCleanup(gossip.unregister_token, _TOKEN)
 
         self.session = run_tests_assert_success(SampleTest)
-        self.test_ids = [result.test_metadata.id for result in self.session.results.iter_test_results()]
+        self.test_ids = [
+            result.test_metadata.id for result in self.session.results.iter_test_results()]
         self._test_all_run()
         self._test_test_logs_written()
         self._test_session_logs()
@@ -96,13 +118,14 @@ class LoggingTest(TestCase):
         methods = [
             method_name for method_name in dir(SampleTest)
             if method_name.startswith("test")
-            ]
+        ]
         self.assertTrue(methods)
         self.assertEquals(len(self.test_ids), len(methods))
 
     def _test_test_logs_written(self):
         for test_id in self.test_ids:
-            log_path = os.path.join(self.log_path, self.session.id, test_id, "debug.log")
+            log_path = os.path.join(
+                self.log_path, self.session.id, test_id, "debug.log")
             with open(log_path) as f:
                 data = f.read()
             for other_test_id in self.test_ids:
@@ -125,7 +148,9 @@ class LoggingTest(TestCase):
                 assert filename.endswith(".log")
                 filename = os.path.join(path, filename)
                 with open(filename) as f:
-                    assert _silenced_logger.name not in f.read(), "Silenced logs appear in log file {0}".format(filename)
+                    assert _silenced_logger.name not in f.read(
+                    ), "Silenced logs appear in log file {0}".format(filename)
+
 
 class ExtraLoggersTest(TestCase):
 
@@ -143,23 +168,30 @@ class ExtraLoggersTest(TestCase):
                 if test_result.test_id in record.message:
                     break
             else:
-                self.fail("Test id {} does not appear in logger".format(test_result.test_id))
+                self.fail(
+                    "Test id {} does not appear in logger".format(test_result.test_id))
+
 
 class SampleTest(slash.Test):
+
     def test_1(self):
         _mark()
+
     def test_2(self):
         _silenced_logger.error("error")
         _silenced_logger.info("info")
         _silenced_logger.debug("debug")
         _mark()
 
+
 def _mark(text=None):
     if text is None:
         text = slash.context.test_id
     slash.logger.debug(text)
 
+
 class TestLocaltimeLogging(TestCase):
+
     def setUp(self):
         super(TestLocaltimeLogging, self).setUp()
         self.assertFalse(slash.config.root.log.localtime)